Improved event handling in HDC-host proxy classes.
The event-proxy classes parse raw event messages into custom "event-payload" objects, which are stored in a deque (aka "double ended queue") to allow for users to process events in a delayed manner and in the thread of their choice.
Size of the deque can be tuned. When full it will drop oldest event-payloads and avoid memory depletion issues. The deque is thread-safe, thus allows data consumption from any thread.

Users can still register handlers that allow immediate reaction to an arriving event. Documentation of the registration method warns users about typical pitfalls when implementing that kind of event-handlers.