Major clean-up of hdc_device implementation.
Got rid of the messy code-duplicity for the HDC-packetizing, which had grown organically during early concept development of the HDC-Spec.

Now there are just two packetizer implementations:
* HDC_Compose_Packets() takes an HDC-message provided as a single block of contiguous data.
* HDC_Compose_Packets_From_Pieces() takes loose bits and pieces and composes and packetizes a message in a single pass.

It's now clearer which methods compose packets into the TX buffers and which methods are actually to be used as API by the main application.

Also implemented proper API methods for:
* Raising custom events: HDC_Raise_Event()
* Custom replies to FeatureCommands requests: HDC_Reply_From_Pieces()
* Void replies to FeatureCommand requests: HDC_Reply_Void()