First steps towards packaging all python code into "hdcproto"
This is work in progress on issue #7

It has been a painfully learned lesson, that all packaging tools rely on working on a top-folder which corresponds to the package and that all python modules *must* be placed in a dedicated child-folder within.
That was also the reason why absolute import statements were not working, which

The name "hdc" is already taken on PyPi, and therefore I had to come up with something unique and concise: hdcproto
Disregarded using any hyphens nor underscores, because they can cause lots of confusion: package-names should use hyphens, but module-names can only use underscores. The best practice seem to be to just use a simple concise single word name.

The package is not working correctly, yet. But we are getting there...
This is approximately the state of the code of this release https://test.pypi.org/project/hdcproto/0.0.6/

I also had to struggle with PyCharm: had to re-install it and also the Python 3.10 package from the "Microsoft Store"