Major clean-up of python logging in hdcproto.
Now that we seem to have a proper package structure, we can use module loggers in the conventional way and extend those with the hierarchy of proxy classes to produce a highly customizable logging backbone for users.

Also cleaned up some of the logs, but that might need further polishing once we actually use those logs of serious debugging or to demonstrate the internals of HDC.

Work in progress on issue #6.
Pending ToDo: Enrich log records with details about the instance, i.e. connection URL, or any kind of unique-ID to differentiate between devices in situations when multiple ones are connected simultaneously.