Utility method to configure LogLevelThreshold setting on HDC-device and keeping the corresponding python logLevel in sync with it.
Considered whether the LogEventThreshold property-proxy should default to a freshness of infinite, so it's always taken from the cache. But this would be a micro-optimization with risks that are not worth it, i.e. forgetting to invalidate the cached value whenever the device resets, but the proxies stay around.

This is work in progress on issue #6.