hdc_device: Renamed API methods from HDC_Raise_Event_... to just HDC_EvtMsg_...
To keep it similar to how the HDC_CmdReply_...() methods are named.

We can *NOT* name it HDC_EvtReply_..., because it would imply that it has been "requested", which it isn't.
Since there is only one allowed transmission direction for EventMessages, it is OK to just name the API ...EvtMsg...