Major fix of naming: It's PACKET, not PACKAGE, in the context of HDC packetizing.
During initial concept development I was quite lax about this, but it's extremely confusing.
Decided to stick to PACKET, because it is the term typically used in the networking domain.
Note, though, that a HDC-packet is by no means a TCP packet!

Also note that a PACKAGE is how we distribute software  libraries. ;-)