STM32: Packetizer that can deal with data streams.
This is a prerequisite for the device firmware to be able to reply with a dynamically generated JSON representation of its HDC interface.
Implemented it as HDC_Compose_Packets_From_Stream(), which in XXXX still coexists with its predecessor HDC_Compose_Packets_From_Pieces(), but will soon replace it.

This is work in progress on issue #21