STM32: Dynamic generation of JSON representation of the HDC-API implemented on a device.
Turned out to be less painful than expected. :-)

Note how the 200ms timeout of the showcase_meta.py script is not sufficient anymore to transmit the 3635 bytes of the reply and therefore had to be increased.
But also note that it is nevertheless more efficient to obtain the whole information in a single call, than the hundreds of calls that it would have required to enquire each information individually through the now obsolete "introspection" method.

This is work in progress on issue #21