Minor, cosmetic fix of ID of the UTF8 datatype.
This is not important at all. :-)

Had originally chosen 0xFF for no particular reason.
Now using 0xAF because it is a better mnemonic for "AlFanumeric string", but also to free the 0xFF, which we might better reserve for ugly extensions of the addressing scheme.

Also because I am about to introduce a datatype for datatypes, and that will now collate more nicely beneath UTF8 and BLOB at 0xD1