STM32: Renamed descriptor structs and typedefs.
What we have until now been calling a "Descriptor" in the C implementation is in fact what we will continue to consider a descriptor.
But I fixed the word order, so that alphabetic sorting of all tokens will make all descriptors cluster together.

Pure renaming. No behaviour changed.

Work in progress on #24