python: Pure FeatureDescriptor(s) extracted from Service class
Services take care to ensure that their respective descriptor is included in the descriptor of their feature. Note that it's OK if they already were, ;-)
Renamed the descriptor attribute Feature.class to Feature.cls, because "class" being a reserved keyword in Python would have otherwise made it impossible to use kwarg expansion in the constructor of the FeatueDescriptor with a bare IDL-dict.

Work in progress on #24