python: Descriptor name validation.
RegEx is not final.
Note there is a difference between a VALID name and a WELL-FORMED name.
We do not worry at this point about names colliding with reserved keywords in Python, C, C#, ... because we are not yet ready to thing about auto-generated source-code. We'll likely will decorating names with prefixes anyway, so it might not be an issue, after all.